song-vocab/requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
ollama>=0.1.6
instructor>=0.4.8
duckduckgo-search>=4.4
aiosqlite>=0.19.0
pydantic>=1.10.13,<2.0.0
python-dotenv>=1.0.0
httpx>=0.25.2
beautifulsoup4>=4.12.2
pytest>=8.0.0
pytest-asyncio>=0.23.5 